Revert "Merge pull request #30 from Kernel360/feat/#28"
This reverts commit 2f9a7cbfe5211f2217965890338f103c81e0ebe6, reversing
changes made to f0b4669b8a21b6dfe6d5e30c8c8fda2ca5c18f77.
