+++ src/features/manager/pages/ManagerInquiries.tsx
@@ -1,31 +1,26 @@
-import { Fragment, useState, useRef, useEffect } from "react";
-import type { SearchManagerReviews as ManagerReviewType } from "@/features/manager/types/ManagerReviewlType";
+import { Fragment, useEffect, useState, useRef } from "react";
+import { Link } from "react-router-dom";
+import { searchManagerInquiries } from "@/features/manager/api/managerInquiry";
+import type { SearchManagerInquiries as ManagerInquiryType } from "@/features/manager/types/ManagerInquirylType";
 import { isValidDateRange } from "@/shared/utils/validation";
-import { searchManagerReviews } from "@/features/manager/api/managerReview";
-import { REVIEW_PAGE_SIZE } from "@/shared/constants/constants";
 
-export const ManagerReviews = () => {
+const PAGE_SIZE = 10;
+
+export const ManagerInquiries = () => {
   const [fadeKey, setFadeKey] = useState(0);
-  const [reviews, setReviews] = useState<ManagerReviewType[]>([]);
+  const [inquiries, setInquiries] = useState<ManagerInquiryType[]>([]);
   const [total, setTotal] = useState(0);
   const [page, setPage] = useState(0);
   const [fromCreatedAt, setFromCreatedAt] = useState<string>("");
   const [toCreatedAt, setToCreatedAt] = useState<string>(""); 
-  const [ratingOption, setRatingOption] = useState(0); // 평점 검색 조건 (0 = 전체, 1~5 = 해당 점수)
-  const [customerNameKeyword, setCustomerNameKeyword] = useState("");
+  const [replyStatus, setReplyStatus] = useState("");
+  const [titleKeyword, setTitleKeyword] = useState("");
   const [contentKeyword, setContentKeyword] = useState("");
   const fromDateRef = useRef<HTMLInputElement>(null);
 
-  const fetchReviews = (paramsOverride?: Partial<ReturnType<typeof getCurrentParams>>) => {
+  const fetchInquiries = (paramsOverride?: Partial<ReturnType<typeof getCurrentParams>>) => {
     const params = getCurrentParams();
-    const finalParams = {
-      ...params,
-      ...paramsOverride,
-      ratingOption: // 여기서 ratingOption 0이면 제거
-        (paramsOverride?.ratingOption ?? params.ratingOption) === 0
-          ? undefined
-          : (paramsOverride?.ratingOption ?? params.ratingOption),
-    };
+    const finalParams = { ...params, ...paramsOverride };
 
     if (!isValidDateRange(finalParams.fromCreatedAt, finalParams.toCreatedAt)) {
       alert("시작일은 종료일보다 늦을 수 없습니다.");
@@ -33,63 +28,70 @@ export const ManagerReviews = () => {
       return;
     }
 
-    searchManagerReviews(finalParams)
+    searchManagerInquiries(finalParams)
       .then((res) => {
-        setReviews(res.content);
+        setInquiries(res.content);
         setTotal(res.page.totalElements);
         setFadeKey((prev) => prev + 1);
       })
       .catch((err) => {
-        console.error("매니저 리뷰 목록 조회 실패:", err);
+        console.error("문의사항 목록 조회 실패:", err);
       });
   };
 
   const getCurrentParams = () => ({
     fromCreatedAt,
     toCreatedAt,
-    ratingOption,
-    customerNameKeyword,
+    replyStatus,
+    titleKeyword,
     contentKeyword,
     page,
-    size: REVIEW_PAGE_SIZE,
+    size: PAGE_SIZE,
   });
 
   useEffect(() => {
-    fetchReviews();
+    fetchInquiries();
   }, [page]);
 
   const handleSearch = () => {
     setPage(0);
-    fetchReviews({ page: 0 });
+    fetchInquiries({ page: 0 });
   };
 
   const handleReset = () => {
     const resetState = {
       fromCreatedAt: "",
       toCreatedAt: "",
-      ratingOption: 0,
-      customerNameKeyword: "",
+      replyStatus: "",
+      titleKeyword: "",
       contentKeyword: "",
       page: 0,
     };
 
     setFromCreatedAt(resetState.fromCreatedAt);
     setToCreatedAt(resetState.toCreatedAt);
-    setRatingOption(resetState.ratingOption);
-    setCustomerNameKeyword(resetState.customerNameKeyword);
+    setReplyStatus(resetState.replyStatus);
+    setTitleKeyword(resetState.titleKeyword);
     setContentKeyword(resetState.contentKeyword);
     setPage(0);
 
-    fetchReviews(resetState);
+    fetchInquiries(resetState);
   };
 
-  const totalPages = Math.max(Math.ceil(total / REVIEW_PAGE_SIZE), 1);
+  const totalPages = Math.max(Math.ceil(total / PAGE_SIZE), 1);
 
   return (
     <Fragment>
       <div className="flex-1 self-stretch inline-flex flex-col justify-start items-start">
         <div className="self-stretch h-16 px-6 bg-white border-b border-gray-200 inline-flex justify-between items-center">
-          <div className="justify-start text-gray-900 text-xl font-bold font-['Inter'] leading-normal">리뷰 관리</div>
+          <div className="justify-start text-gray-900 text-xl font-bold font-['Inter'] leading-normal">문의사항</div>
+          <Link
+            to="/managers/inquiries/new"
+            className="h-10 px-4 bg-indigo-600 rounded-md flex justify-center items-center gap-2 cursor-pointer hover:bg-indigo-700 transition"
+          >
+            <span className="material-symbols-outlined text-white">add</span>
+            <span className="text-white text-sm font-semibold font-['Inter'] leading-none">문의사항 등록</span>
+          </Link>
         </div>
         
         <div className="self-stretch p-6 flex flex-col justify-start items-start gap-6">
@@ -104,7 +106,7 @@ export const ManagerReviews = () => {
             <div className="self-stretch flex flex-col justify-start items-start gap-4">
               <div className="self-stretch inline-flex justify-start items-start gap-4">
                 <div className="flex-1 inline-flex flex-col justify-start items-start gap-2">
-                  <div className="self-stretch justify-start text-slate-700 text-sm font-medium font-['Inter'] leading-none">리뷰 작성일</div>
+                  <div className="self-stretch justify-start text-slate-700 text-sm font-medium font-['Inter'] leading-none">등록일</div>
                   <div className="self-stretch inline-flex justify-start items-center gap-2">
                     <input
                         type="date"
@@ -123,40 +125,37 @@ export const ManagerReviews = () => {
                     </div>
                   </div>
                 <div className="flex-1 inline-flex flex-col justify-start items-start gap-2">
-                  <div className="self-stretch justify-start text-slate-700 text-sm font-medium font-['Inter'] leading-none">평점</div>
+                  <div className="self-stretch justify-start text-slate-700 text-sm font-medium font-['Inter'] leading-none">답변 상태</div>
                   <select
-                    value={ratingOption}
-                    onChange={(e) => setRatingOption(Number(e.target.value))}
+                    value={replyStatus}
+                    onChange={(e) => setReplyStatus(e.target.value)}
                     className="w-full h-12 px-4 bg-slate-50 rounded-lg border border-slate-200 text-slate-700 text-sm focus:outline-indigo-500"
                   >
-                    <option value={0}>전체</option>
-                    <option value={5}>5점</option>
-                    <option value={4}>4점</option>
-                    <option value={3}>3점</option>
-                    <option value={2}>2점</option>
-                    <option value={1}>1점</option>
+                    <option value="">전체</option>
+                    <option value="PENDING">답변 대기</option>
+                    <option value="ANSWERED">답변 완료</option>
                   </select>
                 </div>
               </div>
               <div className="self-stretch inline-flex justify-start items-start gap-4">
                 <div className="flex-1 inline-flex flex-col justify-start items-start gap-2">
-                  <div className="self-stretch justify-start text-slate-700 text-sm font-medium font-['Inter'] leading-none">고객명</div>
+                  <div className="self-stretch justify-start text-slate-700 text-sm font-medium font-['Inter'] leading-none">제목</div>
                   <div className="self-stretch h-12 px-4 bg-slate-50 rounded-lg outline outline-1 outline-offset-[-1px] outline-slate-200 inline-flex justify-start items-center">
                     <input
-                      value={customerNameKeyword}
-                      onChange={(e) => setCustomerNameKeyword(e.target.value)}
-                      placeholder="고객명 입력"
+                      value={titleKeyword}
+                      onChange={(e) => setTitleKeyword(e.target.value)}
+                      placeholder="제목 검색"
                       className="w-full text-sm text-slate-700 placeholder:text-slate-400 bg-transparent focus:outline-none"
                     />
                   </div>
                 </div>
                 <div className="flex-1 inline-flex flex-col justify-start items-start gap-2">
-                  <div className="self-stretch justify-start text-slate-700 text-sm font-medium font-['Inter'] leading-none">리뷰 내용</div>
+                  <div className="self-stretch justify-start text-slate-700 text-sm font-medium font-['Inter'] leading-none">내용</div>
                   <div className="self-stretch h-12 px-4 bg-slate-50 rounded-lg outline outline-1 outline-offset-[-1px] outline-slate-200 inline-flex justify-start items-center">
                     <input
                       value={contentKeyword}
                       onChange={(e) => setContentKeyword(e.target.value)}
-                      placeholder="리뷰 내용 검색"
+                      placeholder="내용 검색"
                       className="w-full text-sm text-slate-700 placeholder:text-slate-400 bg-transparent focus:outline-none"
                     />
                   </div>
@@ -181,67 +180,51 @@ export const ManagerReviews = () => {
           </form>
 
 
-
-          <div className="w-full p-6 bg-white rounded-xl shadow-[0px_2px_12px_0px_rgba(0,0,0,0.04)] inline-flex flex-col justify-start items-start gap-6">
-            <div className="self-stretch inline-flex justify-between items-center">
-              <div className="justify-start text-slate-800 text-lg font-semibold font-['Inter'] leading-snug">리뷰 내역</div>
+          <div className="self-stretch p-6 bg-white rounded-xl shadow-[0px_2px_12px_0px_rgba(0,0,0,0.04)] flex flex-col justify-start items-start">
+            <div className="self-stretch inline-flex justify-between items-center pb-4">
+              <div className="justify-start text-slate-800 text-lg font-semibold font-['Inter'] leading-snug">문의사항 내역</div>
               <div className="justify-start text-slate-500 text-sm font-normal font-['Inter'] leading-none">총 {total}건</div>
             </div>
+            <div className="self-stretch h-12 px-4 bg-slate-50 border-b border-slate-200 inline-flex justify-start items-center">
+              <div className="flex-1 flex justify-center items-center">
+                <div className="flex-1 flex justify-center items-center">
+                  <div className="flex-1 flex justify-center items-center gap-4">
+                    <div className="w-45 text-center text-sm font-semibold text-slate-700 font-semibold font-['Inter'] leading-none">번호</div>
+                    <div className="flex-1 text-center text-sm font-semibold text-slate-700 font-semibold font-['Inter'] leading-none">제목</div>
+                    <div className="w-60 text-center text-sm font-semibold text-slate-700 font-semibold font-['Inter'] leading-none">작성일시</div>
+                    <div className="w-60 text-center text-sm font-semibold text-slate-700 font-semibold font-['Inter'] leading-none">답변 상태</div>
+                  </div>
+                </div>
+              </div>
+            </div>
+
+
             <div key={fadeKey} className="w-full fade-in">
-              {reviews.length === 0 ? (
+              { inquiries.length === 0 ? (
                 <div className="self-stretch h-16 px-4 border-b border-slate-200 flex items-center text-center">
-                  <div className="w-full text-sm text-slate-500">조회된 리뷰가 없습니다.</div>
+                  <div className="w-full text-sm text-slate-500">조회된 문의사항이 없습니다.</div>
                 </div>
               ) : (
-                reviews.map((review) => (
-                  <div
-                    key={review.reviewId}
-                    className="self-stretch p-6 bg-slate-50 rounded-lg flex flex-col justify-start items-start gap-4 mb-4"
-                  >
-                    {/* 상단: 고객명, 날짜, 서비스명, 별점 */}
-                    <div className="self-stretch inline-flex justify-between items-center">
-                      <div className="flex justify-start items-center gap-3">
-                        <div className="inline-flex flex-col justify-center items-start">
-                          <div className="text-slate-800 text-base font-semibold leading-tight">
-                            {review.authorName}
-                          </div>
-                          <div className="text-slate-500 text-sm font-normal leading-none">
-                            {review.createdAt}
-                          </div>
-                        </div>
-                      </div>
-                      <div className="flex justify-start items-center gap-1.5">
-                        <div className="h-7 px-3 bg-sky-100 rounded-2xl flex justify-center items-center">
-                          <div className="text-sky-900 text-sm font-medium leading-none">
-                            {review.serviceName}
-                          </div>
-                        </div>
-                        <div className="flex justify-end items-center gap-1">
-                          {[1, 2, 3, 4, 5].map((i) => (
-                            <span
-                              key={i}
-                              className={`material-icons-outlined text-base ${
-                                review.rating >= i ? 'text-yellow-400' : 'text-slate-300'
-                              }`}
-                            >
-                              star
-                            </span>
-                          ))}
-                          <div className="text-slate-700 text-sm font-semibold leading-none ml-1">
-                            {review.rating.toFixed(1)}
-                          </div>
+                inquiries.map((inquiry) => (
+                  <Link 
+                    key={inquiry.inquiryId}
+                    to={`/managers/inquiries/${inquiry.inquiryId}`}
+                    className="self-stretch h-16 px-4 border-b border-slate-200 flex items-center text-center gap-4">
+                    <div className="w-45 text-center text-sm text-slate-700 font-medium font-['Inter'] leading-none">{inquiry.inquiryId}</div>
+                    <div className="flex-1 flex items-center text-sm text-slate-700 text-left font-medium font-['Inter'] leading-none">{inquiry.title}</div>
+                    <div className="w-60 text-center text-sm text-slate-700 font-medium font-['Inter'] leading-none">{inquiry.createdAt}</div>
+                    <div className="w-60 text-center flex justify-center">
+                      <div className={`h-7 px-3 rounded-2xl flex items-center font-medium font-['Inter'] leading-none ${inquiry.isReplied ? 'bg-green-100' : 'bg-yellow-100'}`}>
+                        <div className={`text-sm font-medium ${inquiry.isReplied ? 'text-green-800' : 'text-yellow-800'}`}>
+                          {inquiry.isReplied ? '답변 완료' : '답변 대기'}
                         </div>
                       </div>
                     </div>
-
-                    {/* 하단: 리뷰 내용 */}
-                    <div className="self-stretch text-slate-700 text-base font-normal leading-tight">
-                      {review.content}
-                    </div>
-                  </div>
+                  </Link>
                 ))
               )}
             </div>
+
             {/* 페이징 */}
             <div className="self-stretch flex justify-center gap-1 pt-4">
               <div
@@ -268,7 +251,7 @@ export const ManagerReviews = () => {
                 <div className="text-sm font-medium font-['Inter'] leading-none">다음</div>
               </div>
             </div>
-          </div> 
+          </div>
 
         </div>
       </div>
