+++ src/features/manager/pages/ManagerInquiryDetail.tsx
