--- src/shared/constants/constants.ts
@@ -1,6 +1,0 @@
-// 리스트 페이지네이션 시 기본 페이지 크기 (1페이지당 항목 수)
-export const DEFAULT_PAGE_SIZE = 10;
-
-
-// 리뷰 리스트 페이지네이션 시 기본 페이지 크기 (1페이지당 항목 수)
-export const REVIEW_PAGE_SIZE = 5;
